+++ modules.admin/admin-entities/src/main/java/cn/com/rexen/admin/entities/OrganizationBean.java
@@ -21,9 +21,9 @@ public class OrganizationBean extends PersistentEntity {
     private String name; //机构名称
     private String code; //机构代码
     private String centerCode; //中心代码
-    private int isLeaf; //是否是叶子节点
+    private long isLeaf; //是否是叶子节点
     private long parentId; //父机构
-    private int areaId;  // 归属区域
+    private long areaId;  // 归属区域
 
 
     public long getParentId() {
@@ -34,7 +34,7 @@ public class OrganizationBean extends PersistentEntity {
         this.parentId = parentId;
     }
 
-    public int getAreaId() {
+    public long getAreaId() {
         return areaId;
     }
 
@@ -66,7 +66,7 @@ public class OrganizationBean extends PersistentEntity {
         this.centerCode = centerCode;
     }
 
-    public int getIsLeaf() {
+    public long getIsLeaf() {
         return isLeaf;
     }
